# src/services/report/routes/v1/public/doc/openapi.json
{
  "openapi": "3.0.0",
...
  "info": {
    "title": "ezReeport API",
    "version": "1.0",
    "contact": {
      "name": "ezTeam",
      "url": "https://github.com/ezpaarse-project",
      "email": "ezpaarse@couperin.org"
    },
    "license": {
      "name": "CeCILL",
      "url": "https://github.com/ezpaarse-project/ezreeport/blob/master/LICENSE.txt"
    },
    "description": "Reporting service"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Direct"
    },
    {
      "url": "/report/api/v1",
      "description": "ezMESURE"
    }
  ],
  "paths": {
...
    "/crons": {
      "get": {
        "summary": "Get all crons",
        "tags": [
          "cron"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": [
                        {
                          "name": "daily-report",
                          "running": true,
                          "lastRun": "2022-11-16T07:21:00.405Z",
                          "nextRun": "2022-11-16T08:22:00.000+01:00"
                        }
                      ]
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Cron"
                      }
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": [
                        {
                          "name": "generateReports",
                          "running": true,
                          "nextRun": "2022-11-16T08:22:00.000+01:00"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "crons-get",
        "security": [
          {
            "User Token": []
          }
        ],
        "description": "Get all crons.\n\nNeeds `general.crons-get` (see `/me/permissions` for more info)."
      }
    },
    "/crons/{cron}": {
      "get": {
        "summary": "Get specific cron",
        "tags": [
          "cron"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "daily-report",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000+01:00"
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/Cron"
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "generateReports",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000Z",
                        "lastRun": "2022-11-16T09:09:59.248Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "crons-get-cron",
        "description": "Get specific crons.\n\nNeeds `general.crons-get-cron` (see `/me/permissions` for more info).",
        "security": [
          {
            "User Token": []
          }
        ]
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "cron",
          "in": "path",
          "required": true,
          "description": "Cron name"
        }
      ]
    },
    "/crons/{cron}/start": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "cron",
          "in": "path",
          "required": true,
          "description": "Cron name"
        }
      ],
      "put": {
        "summary": "Start cron",
        "operationId": "crons-put-cron-start",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "daily-report",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000+01:00"
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/Cron"
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "generateReports",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000Z",
                        "lastRun": "2022-11-16T09:09:59.248Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "description": "Start specific cron.\n\nNeeds `general.crons-put-cron-start` (see `/me/permissions` for more info).",
        "tags": [
          "cron"
        ],
        "security": [
          {
            "User Token": []
          }
        ]
      }
    },
    "/crons/{cron}/stop": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "cron",
          "in": "path",
          "required": true,
          "description": "Cron name"
        }
      ],
      "put": {
        "summary": "Stop cron",
        "operationId": "crons-put-cron-stop",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "daily-report",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000+01:00"
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/Cron"
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "generateReports",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000Z",
                        "lastRun": "2022-11-16T09:09:59.248Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "description": "Stop specific cron.\n\nNeeds `general.crons-put-cron-stop` (see `/me/permissions` for more info).",
        "tags": [
          "cron"
        ],
        "security": [
          {
            "User Token": []
          }
        ]
      }
    },
    "/crons/{cron}/force": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "cron",
          "in": "path",
          "required": true,
          "description": "Cron name"
        }
      ],
      "post": {
        "summary": "Force cron",
        "operationId": "crons-post-cron-force",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "example-1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "daily-report",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000+01:00"
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/Cron"
                    }
                  }
                },
                "examples": {
                  "Example": {
                    "value": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "generateReports",
                        "running": true,
                        "nextRun": "2022-11-16T09:10:00.000Z",
                        "lastRun": "2022-11-16T09:09:59.248Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "description": "Force specific cron to run.\n\nNeeds `general.crons-post-cron-force` (see `/me/permissions` for more info).",
        "tags": [
          "cron"
        ],
        "security": [
          {
            "User Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cron": {
        "type": "object",
        "x-examples": {
          "Example": {
            "name": "daily-report",
            "running": true,
            "nextRun": "2022-11-16T09:10:00.000Z",
            "lastRun": "2022-11-16T09:09:59.000Z"
          }
        },
        "description": "Cron job info",
        "properties": {
          "name": {
            "type": "string"
          },
          "running": {
            "type": "boolean"
          },
          "nextRun": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastRun": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "User Token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Used by user to interact with service"
      },
      "API Key": {
        "name": "X-API-Key",
        "type": "apiKey",
        "in": "header",
        "description": "Used by linked application to manage service"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "JWT malformed: invalid signature"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'GET /users/' requires auth"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'...' requires auth"
                  }
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User 'user' doesn't have the rights to acces to 'GET /users/'"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User '...' doesn't have the rights to access to '...'"
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 404,
                    "message": "Not Found"
                  },
                  "content": {
                    "message": "User 'testa' not found"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 404,
                    "message": "Not Found"
                  },
                  "content": {
                    "message": "... '...' not found"
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "Unexpected token v in JSON at position 0"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}